Big refactor of libraries to escape the need for internal imports and circular dependencies
