+++ tests/test_server_mon.py
@@ -1,11 +1,8 @@
 import datetime
-from unittest.mock import patch, sentinel, Mock
+from unittest.mock import Mock
 
-import requests
-from requests.exceptions import SSLError
-
-from generalised_functions import ResultHolder, ErrorHandler, DAY_TIME_FMT
-from server_mon import interrog_routine
+from generalised_functions import ResultHolder
+from indie_gen_funcs import ErrorHandler
 
 
 def configure_mock_get(mock_get, mock_http_response_time):
@@ -27,79 +24,6 @@ def get_ave_max_latencies(latencies):
     return ave_latency, max_latency
 
 
-@patch("server_mon.CheckResult", autospec=True)
-@patch.object(requests, "get", autospec=True)
-@patch("server_mon.SSHInterrogator.do_queries", autospec=True)
-def test_interrog_routine(mock_queries, mock_get, mock_check_result):
-    mock_http_response_time = 0.0421
-    sample_latencies = ['16.0', '15.0', '18.0', '16.0']
-    configure_mock_get(mock_get, mock_http_response_time)
-    err_handler, result_holder = get_interrog_mock_args()
-    ave_latency, max_latency = get_ave_max_latencies(sample_latencies)
-    interrog_routine(
-        err_handler,
-        {"home_page": sentinel.home_page},
-        result_holder, sentinel.ipv4, sample_latencies)
-    mock_check_result.assert_called_once_with(
-        result_holder.time.strftime(DAY_TIME_FMT), sentinel.ipv4,
-        ave_latency, max_latency,
-        str(int(round(1000 * mock_http_response_time))),
-        str(mock_get.return_value.status_code), None, None,
-        None, None, None, None)
-    mock_queries.assert_called_once()
-    mock_get.assert_called_once_with(sentinel.home_page, timeout=5, verify=True)
-
-
-@patch("server_mon.CheckResult", autospec=True)
-@patch.object(requests, "get", autospec=True)
-@patch("server_mon.SSHInterrogator.do_queries", autospec=True)
-def test_self_sign_interrog_routine(mock_queries, mock_get, mock_check_result):
-    mock_http_response_time = 0.0421
-    sample_latencies = ['16.0', '15.0', '18.0', '16.0']
-    configure_mock_get(mock_get, mock_http_response_time)
-    err_handler, result_holder = get_interrog_mock_args()
-    interrog_routine(
-        err_handler,
-        {"home_page": sentinel.home_page, "verify": sentinel.vrfy_path},
-        result_holder, sentinel.ipv4, sample_latencies)
-    mock_check_result.assert_called_once()
-    mock_queries.assert_called_once()
-    mock_get.assert_called_once_with(
-        sentinel.home_page, timeout=5, verify=sentinel.vrfy_path)
-
-
-@patch("server_mon.CheckResult", autospec=True)
-@patch.object(requests, "get", autospec=True, side_effect=[
-    SSLError(), Mock(ok=True, status_code=200, elapsed=Mock(total_seconds=lambda:200))])
-@patch("server_mon.SSHInterrogator.do_queries", autospec=True)
-def test_interrog_routine_when_cert_expired(mock_queries, mock_get, mock_check_result):
-    sample_latencies = ['16.0', '15.0', '18.0', '16.0']
-    # configure_mock_get(mock_get, mock_http_response_time)
-    err_handler, result_holder = get_interrog_mock_args()
-    interrog_routine(
-        err_handler,
-        {"home_page": sentinel.home_page},
-        result_holder, sentinel.ipv4, sample_latencies)
-    mock_check_result.assert_called_once()
-    assert mock_get.call_count == 2
-    mock_queries.assert_called_once()
-    err_handler.append.assert_called_once()
-
-
-@patch("server_mon.CheckResult", autospec=True)
-@patch.object(requests, "get", autospec=True, side_effect=requests.exceptions.ConnectionError())
-@patch("server_mon.SSHInterrogator.do_queries", autospec=True)
-def test_interrog_routine_when_server_unresponsive(mock_queries, mock_get, mock_check_result):
-    sample_latencies = ['16.0', '15.0', '18.0', '16.0']
-    err_handler, result_holder = get_interrog_mock_args()
-    interrog_routine(
-        err_handler,
-        {"home_page": sentinel.home_page},
-        result_holder, sentinel.ipv4, sample_latencies)
-    mock_check_result.assert_not_called()
-    mock_get.assert_called_once()
-    mock_queries.assert_not_called()
-    err_handler.append.assert_called_once_with(mock_get.side_effect)
